Added search bar to admin dashboard assay type settings table (#134). Added DataTable.js
