+++ client/package.json
@@ -79,7 +79,7 @@
     "react-router-dom": "6.3.0",
     "react-select": "5.3.1",
     "react-syntax-highlighter": "15.5.0",
-    "react-table": "7.7.0",
+    "react-table": "^7.7.0",
     "react-transition-group": "^4.4.2",
     "sanitize-html": "^2.10.0",
     "sweetalert": "^2.1.2",
